Allow config keys which don't have a period to be overridden
We get the original map from viper config. And then we populate the map
with those keys which do not have a period in them.
After we have unmarshalled the original values, we restore the original
values which did not have periods.

This allows us to atleast override values which don't have period in them.

Not a complete fix, because that needs to be done from within viper itself.
But it fixes the most common cases which don't have period in them.

Fixes #13963